CLI Support for dynamic index names
This patch introduces an equivalent dynamic index name support in the CLI
version of pino-elastic search via a `-r` CLI parameter.

-r is analogous to node's require flag in that it allows users of the
commandline to specify a module require path that contains a factory
function which produces an index in exactly the same fashion as what's
already there in the _usage as a module_ section of the documentation.

Usage:

Given I have a `factory.js` file in the current directory with the following
contents:

```js
module.exports = function (date) {
  // return a date here
}
```

And I use pino-elasticsearch on the commandline according to the following format:

```
cat log | pino-elasticsearch -r ./factory
```

Then I will have the index generated as per the function in the `./factory` module.